sravani-internship-2025/requirements.txt
# Core packages
PyYAML>=6.0
pandas>=1.5
tqdm>=4.64
python-dotenv>=1.0
ratelimit>=2.2

# Model APIs
openai>=1.3
transformers>=4.39
accelerate>=0.26
datasets>=2.14

# LangChain integrations
langchain_openai>=0.1
langchain_ollama>=0.1


scikit-learn>=1.2
sentencepiece>=0.1.99 
protobuf<4.0 

# Jupyter environment
jupyterlab>=4.0
ipykernel>=6.0
vllm